Update simulations script for ML classifier (#113)
* add the ML classifier to the simulations

* update the sample outputs to include the ML classifier

* add sample input file, informed by datadog logs

* update the simulations script to run the production and ML classifiers against the inputs pulled from datadog

* update the outputs to show results of running the production and ML classifiers against the inputs pulled from datadog

* update README and fix filepath

* attempt to improve format for github UI

* Update README.md

* update to use onnx model

* update unit tests

* update code and test to reflect we want only the first element of what the ort session returns

* update the outputs to show results of running the production and ML classifiers against a larger input file

the input file was derived from datadog logs, however it is NOT representative of all traffic and accounts only for cases where the csv lookup found a match.

* use csv.writer() instead of generic file.write()
already in use in other parts of the script. benefits of handling cases where a token contains a ',' (in which case the token should be wrapped in quotes)

* update the outputs to show results of running the production and ML classifiers against a larger file
the input file was derived from datadog logs, however it is NOT representative of all traffic and accounts only for cases where the csv lookup found a match.
bug fix: aggregate file correctly wraps in quotes any tokens that contain commas

* update simulations to output classification name (to supplement classification code)
why: more readable than codes, and can be easily removed programmatically as needed

* update simulation outputs to include classification names (in addition to classification codes)

* fix lint (lines too long)

* fix test setup

* add dependency on onnxruntime
(library for Open Neural Network Exchange (ONNX) models)

* add dependency on onnx

* clean/process the inputs as part of invoking the ML classifier

* set return object to have a default value

* return boolean instead of None

* create subdir 'extracts' for the samples from datadog

* modify behavior for if no ML model is loaded
until we are able to deploy the ML model, let's use a placeholder for its predictions, so that we can visualize the logging

* be specific about what exception type we want to handle

* log a warning if we're not able to get a classification name, and use an empty string instead of none as a fallback (to avoid overlap with keyword)

* update logging. wrap lines that are too long, and use warning() instead of warn()
(latter is deprecated)

---------

Co-authored-by: Lisa Chung <943046+lisac@users.noreply.github.com>